requirements.txt
streamlit
pdfplumber
python-docx
sentence-transformers
scikit-learn
requests
groq
python-dotenv
torch